Savestate: Highlight the Latest Save State
This commit shows a (partial) solution to issue #2913.

It'll add the "<== " string to the right of the name of the latest save state slot, as highlight/marker/tag.
The 'highlight' will be updated when (as example) the iso file is mounted or re-mounted, BUT NOT when a new save state is created. 

Of course, this isn't a good/ideal solution, as right now I don't have enough time to explore all the codes. 
But, at least it's enough to show what I want.